Add relative path option to ParseOptions
Using relative paths in config files today gives some odd behavior.  It
seems to attempt to resolve a relative path based on the $CWD of the
executable being run, but this does not always work.  This behavior is
not ideal when moving config files between experiments because it means
that all config files used must be updated for any paths that may have
changed.  This patch introduces a new field in the ParseOptions class
that allows the user to specify the "anchor" directory that will be
used to resolve any relative paths encountered.

Signed-off-by: Eric B Munson <eric@cobaltspeech.com>