Add install target and libs only build
For use in rubik, we do not need to go through and build all the binary
tools for kaldi and building the libraries only is significantly
faster.  Make a target that allows the building of only the libraries.

Also, add targets to install everything or only the libraries.  This
requires the addition of a new configure variable, prefix, which
defaults to /usr/local.

Signed-off-by: Eric B Munson <eric@cobaltspeech.com>