Update android build script
In configuring the kaldi configure script and build system to support
our android build chain in Jenkins, I broke the
install_android_openblas.sh script.  This should make the changes needed
to work with the new configure options.  As a bonus, we now have targets
for building and installing only the kaldi libraries.

Signed-off-by: Eric B Munson <eric@cobaltspeech.com>